Third release: containers fixed and Java compatibility added.

+++ sort/BaseSorter.h
@@ -6,12 +6,11 @@
 template<typename T>
 class BaseSorter : public ISorter<T> {
 protected:
-    void swap(MutableSequence<T>* seq, int i, int j) {
+    void swap(MutableSequence<T>* seq, size_t i, size_t j) {
         T temp = seq->get(i);
         seq->set(i, seq->get(j));
         seq->set(j, temp);
     }
-
 };
 
 #endif
